requirements.txt
streamlit==1.29.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
pytest==7.4.3